Deletes deprecated files
