--- composer.json
@@ -1,5 +1,0 @@
-{
-  "require": {
-    "slim/slim": "^3.9"
-  }
-}
